Added decorator for HTTP verifier method
