+++ src/rest_api/app/app.py
@@ -1,486 +1,41 @@
-from flask import Flask, request, jsonify
+from flask import Flask, request, jsonify, make_response
+from service_registry import service_registry
+from validate_http import method_required
+from get_service import *
+from post_service import *
+from put_service import *
+
 
 app = Flask(__name__)
 
-def checksum_validation(params):
-    """
-    Validate the checksum of a file.
-
-    Args:
-        params (dict): Dictionary containing the parameters for validation, including 'checksumType', 'fileId', and 'expectedChecksum'.
-
-    Returns:
-        dict: Result of the checksum validation including the service name, parameters, and validation result.
-
-    """
-    checksum_type = params.get('checksumType')
-    file_id = params.get('fileId')
-    expected_checksum = params.get('expectedChecksum')
-    
-    def validate_checksum(checksum_type, file_id, expected_checksum):
-        calculated_checksum = "ABCDEF1234567890"
-        return calculated_checksum == expected_checksum
-
-    is_valid = validate_checksum(checksum_type, file_id, expected_checksum)
-
-    result = {
-        "service": "checksum_validation",
-        "params": params,
-        "is_valid": is_valid
-    }
-    return result
-
-def data_validation(params):
-    """
-    Validate data according to specified rules.
-
-    Args:
-        params (dict): Dictionary containing the parameters for validation, including 'schemaVersion', 'dataId', and 'validationRules'.
-
-    Returns:
-        dict: Result of the data validation including the service name, parameters, and validation result.
-
-    """
-    schema_version = params.get('schemaVersion')
-    data_id = params.get('dataId')
-    validation_rules = params.get('validationRules')
-    
-    def validate_data(schema_version, data_id, validation_rules):
-        validation_passed = True
-        for rule in validation_rules:
-            if rule == "rule1":
-                validation_passed = validation_passed and True
-            elif rule == "rule2":
-                validation_passed = validation_passed and True
-        return validation_passed
-
-    is_valid = validate_data(schema_version, data_id, validation_rules)
-
-    result = {
-        "service": "data_validation",
-        "params": params,
-        "is_valid": is_valid
-    }
-    return result
-
-def routine_control(data):
-    """
-    Control a specific routine by starting, stopping, or getting results.
-
-    Args:
-        data (dict): Dictionary containing the control parameters, including 'routineIdentifier', 'controlType', 'routineParameters', and 'timeout'.
-
-    Returns:
-        dict: Result of the routine control including the service name, data, and execution result.
-
-    Relevant UDS Service: Routine Control
-    SID: 0x31
-    """
-    routine_identifier = data.get('routineIdentifier')
-    control_type = data.get('controlType')
-    routine_parameters = data.get('routineParameters')
-    timeout = data.get('timeout')
-
-    def execute_routine_control(routine_identifier, control_type, routine_parameters, timeout):
-        if control_type == "start":
-            result = f"Routine {routine_identifier} started with parameters {routine_parameters} and timeout {timeout}"
-        elif control_type == "stop":
-            result = f"Routine {routine_identifier} stopped"
-        elif control_type == "result":
-            result = f"Result for routine {routine_identifier}: Success"
-        else:
-            result = "Invalid control type"
-        return result
-
-    result = execute_routine_control(routine_identifier, control_type, routine_parameters, timeout)
-
-    response = {
-        "service": "routine_control",
-        "data": data,
-        "result": result
-    }
-    return response
-
-def write_data_by_identifier(data):
-    """
-    Write data to a specific identifier.
-
-    Args:
-        data (dict): Dictionary containing the write parameters, including 'identifier', 'writeType', and 'dataPayload'.
-
-    Returns:
-        dict: Result of the write operation including the service name, data, and write result.
-
-    Relevant UDS Service: Write Data by Identifier
-    SID: 0x2E
-    """
-    identifier = data.get('identifier')
-    write_type = data.get('writeType')
-    data_payload = data.get('dataPayload')
-
-    address = data_payload.get('address')
-    data_to_write = data_payload.get('data')
-    length = data_payload.get('length')
-
-    def perform_write_operation(identifier, write_type, address, data, length):
-        result = f"Data {data} written to address {address} with length {length} as {write_type} operation using identifier {identifier}"
-        return result
-
-    result = perform_write_operation(identifier, write_type, address, data_to_write, length)
-
-    response = {
-        "service": "write_data_by_identifier",
-        "data": data,
-        "result": result
-    }
-    return response
-
-def read_data_by_identifier(params):
-    """
-    Read data by a specific identifier.
-
-    Args:
-        params (dict): Dictionary containing the parameters for reading, including 'data_format', 'timeout', 'identifier', and 'encryption'.
-
-    Returns:
-        dict: Details of the read operation including the service name and parameters.
-
-    Relevant UDS Service: Read Data by Identifier
-    SID: 0x22
-    """
-    data_format = params.get('data_format', 'dummy')
-    timeout = params.get('timeout', 5)  
-    identifier = params.get('identifier')
-    encryption = params.get('encryption', False)  
-
-    query_params = {
-        "data_Format": data_format,
-        "timeout": timeout,
-        "identifier": identifier,
-        "encryption": encryption
-    }
-
-    result = {
-        "service": "read_data_by_identifier", 
-        "method": "GET",
-        "query_params": query
-
-    }
-    return result
-
-def request_update_status(params):
-    """
-    Request the update status.
-
-    Args:
-        params (dict): Dictionary containing the parameters for the request, including 'timestamp', 'priority', and 'userContext'.
-
-    Returns:
-        dict: Details of the update status request including the service name and parameters.
-
-    Relevant OTA Service: Request Update Status
-    SID: 0x32
-    """
-    timestamp = params.get('timestamp')
-    priority = params.get('priority')
-    user_context = params.get('userContext')
-    user_id = user_context.get('userId') if user_context else None
-    roles = user_context.get('roles') if user_context else None
-   
-    result = {
-        "timestamp": timestamp,
-        "priority": priority,
-        "user_context": user_context,
-        "user_id": user_id,
-        "roles": roles
-    }
-
-    return result
-
-def authentication(data):
-    """
-    Perform authentication using provided credentials.
-
-    Args:
-        data (dict): Dictionary containing the authentication details, including 'auth_type', 'description', 'payload', 'user', and 'password'.
-
-    Returns:
-        dict: Result of the authentication including the service name and provided details.
-
-    Relevant UDS Service: Authentication
-    SID: 0x29
-    """
-    auth_type = data.get('auth_type')
-    description = data.get('description')
-    payload = data.get('payload')
-    user = payload.get('user')
-    password = payload.get('password')
-    
-    data = {"service": "authentication", "auth_type": auth_type, "description": description, 'payload': payload, 'user': user, 'password': password}
-    return data
-
-def request_download(data):
-    """
-    Request a download operation.
-
-    Args:
-        data (dict): Dictionary containing the download request details, including 'type', 'description', 'payload', 'deviceId', 'url', and 'version'.
-
-    Returns:
-        dict: Details of the download request including the service name and provided data.
-
-    Relevant OTA Service: Request Download
-    SID: 0x34
-    """
-    type_ = data.get('type')
-    description = data.get('description')
-    payload = data.get('payload')
-    device_id = payload.get('deviceId')
-    url = payload.get('url')
-    version = payload.get('version')
-    
-    data = { "service": "request_download",'type': type_, 'description': description, 'payload': payload, 'device_id': device_id, 'url': url, 'version': version}
-    return data
-
-def read_dtc_information(params):
-    """
-    Read Diagnostic Trouble Code (DTC) information.
-
-    Args:
-        params (dict): Dictionary containing the DTC information parameters, including 'DTCname', 'number_of_DTC', and 'related_information'.
-
-    Returns:
-        dict: Details of the DTC information read including the service name and query parameters.
-
-    Relevant UDS Service: Read DTC Information
-    SID: 0x19
-    """
-    dtc_name = params.get('DTCname')
-    DTC_number = params.get('number_of_DTC')
-    related_info = params.get('related_information')
-
-    query_params = {
-        "dtc_name": dtc_name,
-        "DTC_number": DTC_number,
-        "related_info": related_info
-    }
-    result = {
-        "service": "read_dtc_information",
-        "method": "GET",
-        "query_params": query_params
-    }
-    return result
-
-def tester_present(data):
-    """
-    Indicate tester presence in a session.
-
-    Args:
-        data (dict): Dictionary containing the tester presence details, including 'presence_flag', 'requestType', 'period', and 'session'.
-
-    Returns:
-        dict: Details of the tester presence including the service name and provided data.
-
-    Relevant UDS Service: Tester Present
-    SID: 0x3E
-    """
-    presence_flag = data.get('presence_flag')
-    request_type = data.get('requestType')
-    period = data.get('period')
-    session = data.get('session')
-    result = {
-        "presence_flag": presence_flag,
-        "requestType": request_type,
-        "period": period,
-        "session": session,
-    }
-    return {"service": "tester_present", "data": result}
-
-def transfer_data(data):
-    """
-    Transfer data with specified parameters.
-
-    Args:
-        data (dict): Dictionary containing the data transfer details, including 'action', 'block_Sequence', and 'memorylocation'.
-
-    Returns:
-        dict: Details of the data transfer including the service name and provided data.
-
-    Relevant UDS Service: Transfer Data
-    SID: 0x36
-    """
-    action = data.get('action')
-    block_sequence = data.get('block_Sequence')
-    memory_location = data.get('memorylocation')
-    result = {
-        "action": action,
-        "block_Sequence": block_sequence,
-        "memorylocation": memory_location
-    }
-    return {"service": "transfer_data", "data": result}
-
-def clear_diagnostic_information(data):
-    """
-    Clear diagnostic information.
-
-    Args:
-        data (dict): Dictionary containing the diagnostic information to clear, including 'system', 'action', and 'DTC'.
-
-    Returns:
-        dict: Details of the clear diagnostic operation including the service name and provided data.
-
-    Relevant UDS Service: Clear Diagnostic Information
-    SID: 0x14
-    """
-    system = data.get('system')
-    action = data.get('action')
-    DTC = data.get('DTC')
-    return {"service": "clear_diagnostic_information",
-            "data": {
-                "system": system,
-                "action": action,
-                "DTC": DTC,
-            }}
-
-def access_timing_parameters(data):
-    """
-    Access and modify timing parameters.
-
-    Args:
-        data (dict): Dictionary containing the timing parameters, including 'session', 'timingType', 'timingValue', and 'retryCount'.
-
-    Returns:
-        dict: Details of the timing parameter access including the service name and provided data.
-
-    Relevant UDS Service: Access Timing Parameters
-    SID: 0x83
-    """
-    session = data.get('session')
-    timing_type = data.get('timingType')
-    timing_value = data.get('timingValue')
-    retry_count = data.get('retryCount')
-    return {
-        "service": "access_timing_parameters",
-        "data": {
-            "session": session,
-            "timingType": timing_type,
-            "timingValue": timing_value,
-            "retryCount": retry_count
-        }
-    }
-
-def read_memory_by_address(params):
-    """
-    Read memory by address.
-
-    Args:
-        params (dict): Dictionary containing the memory read parameters, including 'address' and 'length'.
-
-    Returns:
-        dict: Details of the memory read operation including the service name, address, length, and read data.
-
-    Relevant UDS Service: Read Memory by Address
-    SID: 0x23
-    """
-    address = params.get('address', '0x1234')
-    length = params.get('length', '16')
-    data = "dummy_data"
-    return {"service": "read_memory_by_address", "address": address, "length": length, "data": data}
-
-def session_diagnostic_control(data):
-    """
-    Control a diagnostic session.
-
-    Args:
-        data (dict): Dictionary containing the session control parameters, including 'session_type' and 'timeout'.
-
-    Returns:
-        dict: Details of the session control operation including the service name and provided data.
-
-    Relevant UDS Service: Session Diagnostic Control
-    SID: 0x10
-    """
-    session_type = data.get('session_type')
-    timeout = data.get('timeout')
-    return {
-        "service": "session_diagnostic_control",
-        "data": {
-            "session_type": session_type,
-            "timeout": timeout,
-        }
-    }
-
-def request_transfer_exit(data):
-    """
-    Request to exit a data transfer session.
-
-    Args:
-        data (dict): Dictionary containing the transfer exit parameters, including 'blockSequenceCounter', 'checksum', and 'transfer_result'.
-
-    Returns:
-        dict: Details of the transfer exit request including the service name and provided data.
-
-    Relevant UDS Service: Request Transfer Exit
-    SID: 0x37
-    """
-    block_sequence_counter = data.get('blockSequenceCounter', '10')
-    checksum = data.get('checksum', 'ABCDEF1234567890')
-    transfer_result = data.get('transferResult', 'success')
-    return {"service": "request_transfer_exit",
-            "data": {
-                "blockSequenceCounter": block_sequence_counter,
-                "checksum": checksum,
-                "transfer_result": transfer_result,
-            }}
-
 @app.route('/api', methods=['GET', 'POST', 'PUT'])
-def handle_request():
+@method_required
+def api_route():
     """
-    Handle API requests for various services.
-
-    Args:
-        None
+    Route to handle API requests for different services. Supports GET, POST, and PUT methods.
 
     Returns:
-        Response: JSON response containing the service result or error message.
+        Response: JSON response containing the result of the service function or an error message.
     """
     service = request.args.get('service')
-    params = request.json if request.method in ['POST', 'PUT'] else request.args
-
-    if service == 'checksum_validation':
-        return jsonify(checksum_validation(params))
-    elif service == 'data_validation':
-        return jsonify(data_validation(params))
-    elif service == 'routine_control':
-        return jsonify(routine_control(params))
-    elif service == 'write_data_by_identifier':
-        return jsonify(write_data_by_identifier(params))
-    elif service == 'read_data_by_identifier':
-        return jsonify(read_data_by_identifier(params))
-    elif service == 'request_update_status':
-        return jsonify(request_update_status(params))
-    elif service == 'authentication':
-        return jsonify(authentication(params))
-    elif service == 'request_download':
-        return jsonify(request_download(params))
-    elif service == 'read_dtc_information':
-        return jsonify(read_dtc_information(params))
-    elif service == 'tester_present':
-        return jsonify(tester_present(params))
-    elif service == 'transfer_data':
-        return jsonify(transfer_data(params))
-    elif service == 'clear_diagnostic_information':
-        return jsonify(clear_diagnostic_information(params))
-    elif service == 'access_timing_parameters':
-        return jsonify(access_timing_parameters(params))
-    elif service == 'read_memory_by_address':
-        return jsonify(read_memory_by_address(params))
-    elif service == 'session_diagnostic_control':
-        return jsonify(session_diagnostic_control(params))
-    elif service == 'request_transfer_exit':
-        return jsonify(request_transfer_exit(params))
+    
+    if service in service_registry:
+        service_info = service_registry[service]
+        service_func = service_info['func']
+        
+        if request.method == 'GET':
+            params = request.args.to_dict()
+            result = service_func(params)
+        elif request.method == 'POST':
+            data = request.json
+            result = service_func(data)
+        elif request.method == 'PUT':
+            data = request.json
+            result = service_func(data)
+            
+        return jsonify(result)
     else:
-        return jsonify({"error": "Invalid service"}), 400
+        return jsonify({"error": "Service not found"}), 404
 
 if __name__ == '__main__':
     app.run(debug=True)
